Refactor annotation class CRUD imports and introduce TileStore for tile operations
