+++ quickannotator/api/v1/annotation/__init__.py
@@ -9,7 +9,7 @@ from typing import List
 import quickannotator.db.models as models
 from ..utils.shared_crud import AnnotationStore
 import geojson
-from quickannotator.api.v1.utils.shared_crud import upsert_gt_tiles
+from quickannotator.db.crud.tile import TileStoreFactory
 
 
 
@@ -91,7 +91,8 @@ class Annotation(MethodView):
         polygons: List[geojson.Polygon] = args['polygons']
         store = AnnotationStore(image_id, annotation_class_id, is_gt=True, in_work_mag=False)
         anns = store.insert_annotations([shape(poly) for poly in polygons])
-        upsert_gt_tiles(image_id, annotation_class_id, {ann.tile_id for ann in anns})
+        tilestore = TileStoreFactory.get_tilestore()
+        tilestore.upsert_gt_tiles(image_id, annotation_class_id, {ann.tile_id for ann in anns})
 
         return anns, 200
