+++ quickannotator/tests/conftest.py
@@ -8,7 +8,8 @@ from quickannotator.db import models
 from quickannotator.api.v1.project.utils import add_project
 from quickannotator.api.v1.image.utils import add_image_by_path
 from quickannotator.api.v1.annotation_class.helper import insert_annotation_class
-from quickannotator.api.v1.utils.shared_crud import upsert_gt_tiles, AnnotationStore
+from quickannotator.api.v1.utils.shared_crud import AnnotationStore
+from quickannotator.db.crud.tile import TileStore, TileStoreFactory
 from quickannotator.constants import TileStatus
 from shapely.geometry import Polygon
 from quickannotator.api.v1.utils.coordinate_space import get_tilespace
@@ -86,7 +87,8 @@ def seed(db_session):   # here db_session is the fixture
                         )
 
     # Add a tile
-    upsert_gt_tiles(
+    tilestore: TileStore = TileStoreFactory.get_tilestore()
+    tilestore.upsert_gt_tiles(
         annotation_class_id=2,
         image_id=1,
         tile_ids=[0]
